% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_grecolatino.R
\name{Tabla_ANOVA}
\alias{Tabla_ANOVA}
\title{Diseño en Cuadro Grecolatino}
\usage{
Tabla_ANOVA(respuesta, tratamiento, observaciones, data)
}
\arguments{
\item{respuesta}{(vector)datos de la variable respuesta}

\item{tratamiento}{(vector) datos de mi variable explicativa}

\item{observaciones}{(vector) datos de mis variables totales}

\item{data}{(dataframe)tabla de datos en formato largo}
}
\value{
crea una tabla con los calculos correspondientes para
el analisis de la varianza
}
\description{
Es una tecnica estadistica que permite controlar el efecto de dos o mas
en un estudio
}
\examples{
\dontrun{
Lectura del archivo de datos
library(readxl)
DATOS_DeCG <- read_excel("C:/Users/Ivonne/Desktop/DATOS.DeCG.xlsx")
View(DATOS_DeCG)
df<-data.frame(DATOS_DeCG)

#Ejecutamos la funcion
Tabla_ANOVA(respuesta = tabla_anova,tratamiento = DatOS_DeCG,observaciones = 25,data = data.frame())
}
}
